Week 2, Day 5 project: Frida Kahlo
"You've been hired to work on a retrospective of Frida Kahlo's work at a major museum. Your job is to put together the audio tour, but in order to do that, you need to create a list of each painting featured in the exhibit, the date it was painted, and its spot in the tour.

Use your knowledge of Python lists to create a master list of each painting, its date, and its audio tour ID.

You will be completing this project on your own computer, using a Jupyter notebook. Follow the steps below to set yourself up for success! If you have any questions, make sure to reach out to your fellow students and Slack moderator."